add new busco rule parserscritp WIP
