+++ atlas/workflow/rules/bin_quality.smk
@@ -42,11 +42,11 @@ rule run_busco:
     output:
         "{sample}/binning/{binner}/bin_quality/busco.tsv",
     params:
-        tmpdir = lambda wc: f"{config['tmpdir']}/busco/{wc.sample}",
+        tmpdir = lambda wc: f"{config['tmpdir']}/busco/{wc.sample}_{wc.binner}",
         busco_download = os.path.join(config["database_dir"], "busco_lineages"),
     conda:
-        "%s/busco.yaml" % CONDAENV
-    threads: config.get("threads", 8),
+        "../envs/busco.yaml"
+    threads: config["threads"]
     log:
         "{sample}/logs/binning/{binner}/busco.log",
     benchmark:
@@ -59,111 +59,102 @@ rule run_busco:
         " -o {params.tmpdir} --download_path {params.busco_download} -c {threads} "
         " &> {log} "
         " ; "
-        " mv {params.tmpdir}/batch_summary.txt {output} "
+        " mv {params.tmpdir}/batch_summary.txt {output} 2>> {log}"
 
 
 ##### checkM  #########
-rule run_checkm_lineage_wf:
+rule run_checkm:
     input:
         touched_output="logs/checkm_init.txt",
         bins=bin_quality_input_folder,  # actualy path to fastas
     output:
-        "{sample}/binning/{binner}/checkm/completeness.tsv",
-        "{sample}/binning/{binner}/checkm/storage/tree/concatenated.fasta",
+        completeness= "{sample}/binning/{binner}/bin_quality/checkm.tsv",
+        taxonomy=     "{sample}/binning/{binner}/bin_quality/checkm_taxonomy.tsv",
     params:
-        output_dir=lambda wc, output: os.path.dirname(output[0]),
-        tmpdir=config["tmpdir"],
+        tmp_output_dir = lambda wc: f"{config['tmpdir']}/checkm/{wc.sample}_{wc.binner}",
+        extension ="fasta"
     conda:
-        "%s/checkm.yaml" % CONDAENV
-    threads: config.get("threads", 1)
+        "../envs/checkm.yaml"
+    threads: config["threads"]
     log:
         "{sample}/logs/binning/{binner}/checkm.log",
+    benchmark:
+        "logs/benchmarks/checkm_lineage_wf/{sample}_{binner}.tsv"
     resources:
         time=config["runtime"]["long"],
         mem=config["large_mem"],
-    benchmark:
-        "logs/benchmarks/checkm_lineage_wf/{sample}_{binner}.tsv"
     shell:
-        """
-        rm -r {params.output_dir}
-        checkm lineage_wf \
-            --file {params.output_dir}/completeness.tsv \
-            --tmpdir {params.tmpdir} \
-            --tab_table \
-            --quiet \
-            --extension fasta \
-            --threads {threads} \
-            {input.bins} \
-            {params.output_dir} &> {log}
-        """
+        " checkm lineage_wf "
+        " --file {params.completeness} "
+        " --tab_table "
+        " --extension {params.extension} "
+        " --threads {threads} "
+        " {input.bins} "
+        " {params.tmp_output_dir} &> {log} "
+        " ;\n"
+        " checkm tree_qa "
+        " {params.tmp_output_dir} "
+        " --out_format 2 "
+        " --file {output.taxonomy} "
+        " --tab_table "
+        " &>> {log}"
 
 
-rule run_checkm_tree_qa:
+## Combine bin stats
+localrules:
+    build_bin_report,
+    combine_bin_stats,
+
+if config['bin_quality_asesser'].lower() == 'checkm':
+
+    rule combine_bin_stats:
+        input:
+            completeness_files=expand(
+                "{sample}/binning/{{binner}}/bin_quality/checkm.tsv", sample=SAMPLES
+            ),
+            taxonomy_files=expand(
+                "{sample}/binning/{{binner}}/bin_quality/checkm_taxonomy.tsv", sample=SAMPLES
+            ),
+        output:
+            bin_table="reports/genomic_bins_{binner}.tsv",
+        params:
+            samples=SAMPLES,
+        log:
+            "logs/binning/combine_stats_{binner}.log",
+        script:
+            "../scripts/combine_checkm.py"
+
+elif config['bin_quality_asesser'].lower() == 'busco':
+
+    rule combine_bin_stats:
+        input:
+            completeness_files=expand(
+                "{sample}/binning/{{binner}}/busco/.tsv", sample=SAMPLES
+            ),
+        output:
+            bin_table="reports/genomic_bins_{binner}.tsv",
+        params:
+            samples=SAMPLES,
+        log:
+            "logs/binning/combine_stats_{binner}.log",
+        script:
+            "../scripts/combine_busco.py"
+
+else:
+
+    raise Exception("'bin_quality_asesser' should be 'busco' or 'checkM' got {bin_quality_asesser}".format(**config))
+
+
+
+
+rule build_bin_report:
     input:
-        tree="{checkmfolder}/completeness.tsv",
+        bin_table="reports/genomic_bins_{binner}.tsv",
     output:
-        summary="{checkmfolder}/taxonomy.tsv",
-    params:
-        tree_dir=lambda wc, input: os.path.dirname(input.tree),
+        report="reports/bin_report_{binner}.html",
     conda:
-        "%s/checkm.yaml" % CONDAENV
-    threads: 1
-    resources:
-        mem=config["simplejob_mem"],
-        time=config["runtime"]["simplejob"],
-    shell:
-        """
-        checkm tree_qa \
-           {params.tree_dir} \
-           --out_format 2 \
-           --file {output.summary}\
-           --tab_table
-
-        """
-
-
-rule checkm_tetra:
-    input:
-        contigs=BINNING_CONTIGS,
-    output:
-        "{sample}/binning/{binner}/checkm/tetranucleotides.txt",
+        "%s/report.yaml" % CONDAENV
     log:
-        "{sample}/logs/binning/{binner}/checkm/tetra.txt",
-    conda:
-        "%s/checkm.yaml" % CONDAENV
-    threads: config.get("threads", 8)
-    shell:
-        """
-        checkm tetra \
-        --threads {threads} \
-        {input.contigs} {output} 2> {log}
-        """
-
-
-rule checkm_outliers:
-    input:
-        tetra="{sample}/binning/{binner}/checkm/tetranucleotides.txt",
-        bin_folder="{sample}/binning/{binner}/bins",
-        checkm="{sample}/binning/{binner}/checkm/completeness.tsv",
-    params:
-        checkm_folder=lambda wc, input: os.path.dirname(input.checkm),
-        report_type="any",
-        treshold=95,  #reference distribution used to identify outliers; integer between 0 and 100 (default: 95)
-    output:
-        "{sample}/binning/{binner}/checkm/outliers.txt",
-    log:
-        "{sample}/logs/binning/{binner}/checkm/outliers.txt",
-    conda:
-        "%s/checkm.yaml" % CONDAENV
-    threads: config.get("threads", 8)
-    shell:
-        """
-        checkm outliers \
-        --extension fasta \
-        --distributions {params.treshold} \
-        --report_type {params.report_type} \
-        {params.checkm_folder} \
-        {input.bin_folder} \
-        {input.tetra} \
-        {output} 2> {log}
-        """
+        "logs/binning/report_{binner}.log",
+    script:
+        "../report/bin_report.py"
