+++ atlas/workflow/scripts/utils/__init__.py
@@ -1,3 +1,3 @@
-from . import parsers_bbmap, parsers_checkm
+from . import parsers_bbmap, parsers
 from .utils import gen_names_for_range
 from . import taxonomy
