+++ atlas/workflow/scripts/utils/parsers.py
